requirements.txt
agency-swarm==0.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.1
arxiv==2.1.3
asn1crypto==1.5.1
attrs==24.2.0
bcrypt==4.2.1
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.14.0
colorama==0.4.6
cryptography==43.0.3
datamodel-code-generator==0.26.1
decorator==5.1.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docx==0.2.4
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
feedparser==6.0.11
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
genson==1.3.0
google-auth==2.36.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.4
jiter==0.7.0
jsonref==1.1.0
lxml==5.3.0
markd==0.1.20
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.1.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.3
openai==1.51.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pg8000==1.31.2
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
propcache==0.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.10.1
pypdfium2==4.30.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
rsa==4.9
scramp==1.4.5
selenium==4.26.1
selenium-stealth==1.0.6
setuptools==75.6.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.2
termcolor==2.4.0
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.0
wsproto==1.2.0
yarl==1.18.3
